requirements.txt
Bottleneck==1.0.0
Cython==0.20.1post0
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
Pygments==1.6
Sphinx==1.2.2
argparse==1.2.1
astropy==1.2.1
astroscrappy==1.0.3
ccdproc==1.0.1
chardet==2.0.1
colorama==0.2.5
decorator==3.4.0
docutils==0.11
html5lib==0.999
nose==1.3.1
numpy<1.10
progressbar2
pytz==2012c
requests==2.2.1
roman==2.0.0
scipy==0.17.1
six==1.5.2
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
